+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Add #_PURE annotation to frozen namespaces ([#2044](https://github.com/rollup/rollup/pull/2044))
 * Improve sourcemap and tree-shaking performance ([#2052](https://github.com/rollup/rollup/pull/2052))
 * Inline sourcemap lookups and make rollup smaller ([#2055](https://github.com/rollup/rollup/pull/2055))
+* Use updated magic-string types ([#2057](https://github.com/rollup/rollup/pull/2057))
 * [BREAKING] Revert class id preservation from #2025 ([#2048](https://github.com/rollup/rollup/pull/2048))
 * [BREAKING] Refactor missing export plugin hook ([#1987](https://github.com/rollup/rollup/pull/1987))